requirements.txt
ipykernel
langchain
python-dotenv
langchain_community
langchain-text-splitters
bs4
arxiv
pypdf
PyMuPDF
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain-huggingface
hf_xet
faiss-cpu
langchain_chroma
streamlit
langchain-groq
fastapi
uvicorn
langserve
youtube-transcript-api